ScatterPlot: support user defined column, update animate
Adds a functionality to DataStorage to interpret
string data as numerical data which can then be
used by the ScatterPlot visualizer to turn into
locations.

ScatterPlot also exposes UI to choose which
columns in the csv file should map to which
corresponding axis in three-dimensional space.

animate_objects() functions are updated with
the new interpretation of duration first added
in the ObjectVisualizer.